+++ pilot/pkg/networking/grpcgen/grpcgen.go
@@ -15,23 +15,9 @@
 package grpcgen
 
 import (
-	"net"
-	"strconv"
-	"strings"
-
-	cluster "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
-	core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
-	listener "github.com/envoyproxy/go-control-plane/envoy/config/listener/v3"
-	route "github.com/envoyproxy/go-control-plane/envoy/config/route/v3"
-	hcm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
-	envoyauth "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/tls/v3"
-	discovery "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v3"
-
 	"istio.io/istio/pilot/pkg/model"
-	"istio.io/istio/pilot/pkg/networking/util"
 	v3 "istio.io/istio/pilot/pkg/xds/v3"
 	"istio.io/istio/pkg/config/host"
-	"istio.io/pkg/log"
 )
 
 // Support generation of 'ApiListener' LDS responses, used for native support of gRPC.
@@ -79,200 +65,3 @@ func (g *GrpcConfigGenerator) Generate(proxy *model.Proxy, push *model.PushConte
 	return nil, model.DefaultXdsLogDetails, nil
 }
 
-// handleLDSApiType handles a LDS request, returning listeners of ApiListener type.
-// The request may include a list of resource names, using the full_hostname[:port] format to select only
-// specific services.
-func (g *GrpcConfigGenerator) BuildListeners(node *model.Proxy, push *model.PushContext, names []string) model.Resources {
-	resp := model.Resources{}
-
-	filter := map[string]bool{}
-	for _, name := range names {
-		if strings.Contains(name, ":") {
-			n, _, err := net.SplitHostPort(name)
-			if err == nil {
-				name = n
-			}
-		}
-		filter[name] = true
-	}
-
-	for _, el := range node.SidecarScope.EgressListeners {
-		for _, sv := range el.Services() {
-			shost := string(sv.Hostname)
-			if len(filter) > 0 {
-				// DiscReq has a filter - only return services that match
-				if !filter[shost] {
-					continue
-				}
-			}
-			for _, p := range sv.Ports {
-				hp := net.JoinHostPort(shost, strconv.Itoa(p.Port))
-				ll := &listener.Listener{
-					Name: hp,
-				}
-
-				ll.Address = &core.Address{
-					Address: &core.Address_SocketAddress{
-						SocketAddress: &core.SocketAddress{
-							Address: sv.Address,
-							PortSpecifier: &core.SocketAddress_PortValue{
-								PortValue: uint32(p.Port),
-							},
-						},
-					},
-				}
-				hcm := &hcm.HttpConnectionManager{
-					RouteSpecifier: &hcm.HttpConnectionManager_Rds{
-						Rds: &hcm.Rds{
-							ConfigSource: &core.ConfigSource{
-								ConfigSourceSpecifier: &core.ConfigSource_Ads{
-									Ads: &core.AggregatedConfigSource{},
-								},
-							},
-							RouteConfigName: clusterKey(shost, p.Port),
-						},
-					},
-				}
-				hcmAny := util.MessageToAny(hcm)
-				// TODO: for TCP listeners don't generate RDS, but some indication of cluster name.
-				ll.ApiListener = &listener.ApiListener{
-					ApiListener: hcmAny,
-				}
-				resp = append(resp, &discovery.Resource{
-					Name:     hp,
-					Resource: util.MessageToAny(ll),
-				})
-			}
-		}
-	}
-
-	return resp
-}
-
-// Handle a gRPC CDS request, used with the 'ApiListener' style of requests.
-// The main difference is that the request includes Resources.
-func (g *GrpcConfigGenerator) BuildClusters(node *model.Proxy, push *model.PushContext, names []string) model.Resources {
-	resp := model.Resources{}
-	// gRPC doesn't currently support any of the APIs - returning just the expected EDS result.
-	// Since the code is relatively strict - we'll add info as needed.
-	for _, n := range names {
-		hn, portn, err := net.SplitHostPort(n)
-		if err != nil {
-			log.Warn("Failed to parse ", n, " ", err)
-			continue
-		}
-
-		porti, err := strconv.Atoi(portn)
-		if err != nil {
-			log.Warn("Failed to parse ", n, " ", err)
-			continue
-		}
-
-		// SANS associated with this host name.
-		// TODO: apply DestinationRules, etc
-		sans := push.ServiceAccounts[host.Name(hn)][porti]
-
-		// Assumes 'default' name, and credentials/tls/certprovider/pemfile
-
-		tlsC := &envoyauth.UpstreamTlsContext{
-			CommonTlsContext: &envoyauth.CommonTlsContext{
-				TlsCertificateCertificateProviderInstance: &envoyauth.CommonTlsContext_CertificateProviderInstance{
-					InstanceName:    "default",
-					CertificateName: "default",
-				},
-
-				ValidationContextType: &envoyauth.CommonTlsContext_CombinedValidationContext{
-					CombinedValidationContext: &envoyauth.CommonTlsContext_CombinedCertificateValidationContext{
-						ValidationContextCertificateProviderInstance: &envoyauth.CommonTlsContext_CertificateProviderInstance{
-							InstanceName:    "default",
-							CertificateName: "ROOTCA",
-						},
-						DefaultValidationContext: &envoyauth.CertificateValidationContext{
-							MatchSubjectAltNames: util.StringToExactMatch(sans),
-						},
-					},
-				},
-			},
-		}
-
-		rc := &cluster.Cluster{
-			Name:                 clusterKey(hn, porti),
-			ClusterDiscoveryType: &cluster.Cluster_Type{Type: cluster.Cluster_EDS},
-			EdsClusterConfig: &cluster.Cluster_EdsClusterConfig{
-				ServiceName: "outbound|" + portn + "||" + hn,
-				EdsConfig: &core.ConfigSource{
-					ConfigSourceSpecifier: &core.ConfigSource_Ads{
-						Ads: &core.AggregatedConfigSource{},
-					},
-				},
-			},
-			TransportSocket: &core.TransportSocket{
-				Name:       transportSocketName,
-				ConfigType: &core.TransportSocket_TypedConfig{TypedConfig: util.MessageToAny(tlsC)},
-			},
-		}
-		// see grpc/xds/internal/client/xds.go securityConfigFromCluster
-		resp = append(resp, &discovery.Resource{
-			Name:     n,
-			Resource: util.MessageToAny(rc),
-		})
-	}
-	return resp
-}
-
-// handleSplitRDS supports per-VIP routes, as used by GRPC.
-// This mode is indicated by using names containing full host:port instead of just port.
-// Returns true of the request is of this type.
-func (g *GrpcConfigGenerator) BuildHTTPRoutes(node *model.Proxy, push *model.PushContext, routeNames []string) model.Resources {
-	resp := model.Resources{}
-
-	// Currently this mode is only used by GRPC, to extract Cluster for the default
-	// route.
-	for _, n := range routeNames {
-		_, _, hostname, port := model.ParseSubsetKey(n)
-		if hostname == "" || port == 0 {
-			log.Warn("Failed to parse ", n)
-			continue
-		}
-		hn := string(hostname)
-		el := node.SidecarScope.GetEgressListenerForRDS(port, "")
-		// TODO: use VirtualServices instead !
-		// Currently gRPC doesn't support matching the path.
-		svc := el.Services()
-		for _, s := range svc {
-			if s.Hostname.Matches(host.Name(hn)) {
-				// Only generate the required route for grpc. Will need to generate more
-				// as GRPC adds more features.
-				rc := &route.RouteConfiguration{
-					Name: n,
-					VirtualHosts: []*route.VirtualHost{
-						{
-							Name:    hn,
-							Domains: []string{hn, net.JoinHostPort(hn, strconv.Itoa(port))},
-
-							Routes: []*route.Route{
-								{
-									Match: &route.RouteMatch{
-										PathSpecifier: &route.RouteMatch_Prefix{Prefix: ""},
-									},
-									Action: &route.Route_Route{
-										Route: &route.RouteAction{
-											ClusterSpecifier: &route.RouteAction_Cluster{
-												Cluster: n,
-											},
-										},
-									},
-								},
-							},
-						},
-					},
-				}
-				resp = append(resp, &discovery.Resource{
-					Name:     n,
-					Resource: util.MessageToAny(rc),
-				})
-			}
-		}
-	}
-	return resp
-}
